gps-globe: Add GpsgSphere
This implements a ClutterTexture subclass that renders its texture
onto a sphere. The sphere geometry is generated by subdividing an
icosahedron in a similar way to Blender's icosphere.

The sphere has a property to specify whether to paint the texture,
lines showing the geometry or both. The lines are mostly for
debugging.

The texture coordinates need to be implemented more intelligently.
Currently the tx and ty is just a number based on the latitude or
longitude of the vertex. It should use a better map projection and it
needs to cope better at the seam. This might involve duplicating the
vertices so that vertices along the seam can have different texture
coords depending on which side of the seam the triangle it belongs to
is.