gpsg-sphere: Fix texture coordinates along the seam
Some triangles spread along the seam where the texture should wrap
from the left edge to the right edge. We need to duplicate one of the
vertices in each of these triangles so it can use a texture coordinate
less than zero or greater than one without affecting the other
triangles that share the vertex. Otherwise the fragments in the
triangle will vary across the entire texture so the seam will look
wrong.