gpsg-sphere: Add a property to flatten out the sphere
The 'flatness' property takes a value between 0.0 and 1.0 specifies
whether to render the texture as a sphere or a flat image. The paint
method uses a vertex shader to interpolate between the two states if
the value is between 0 and 1. The vertex coords for the flat geometry
are generated from the texture coordinates.