clutter-ply-data: Store colours as guint8 not floats
The data for a vertex is now stored in a byte array rather than an
array of floats. Each property now has a size and the property map
stores byte offsets rather than float indices. That way the color
values can be stored directly as bytes rather than having to expand
them to floats. As there is usually only 3 bytes for a color then this
ends up with a vertex size that is not a multiple of 4 bytes so it
would end up with unaligned accesses if the vertices were packed
tightly. To work around this it adds padding to the size of a vertex
to round it up to 4 bytes.