clutter-ply-model: Add a 'fit-to-allocation' property
When set this property causes the actor to scale the model to take up
the most amount of space within the actor's allocation without
distoring the aspect ratio. This makes it much easier to render small
models. The corresponding code for the viewer example has been greatly
reduced. The property is enabled by default.