pinpoint: refactor: Abstract the renderering code
With this refactoring, we introduce PinPointRenderer and a derivate
structure ClutterRenderer. Those structures are the only one holding data
for the rendering of the slides, effectively splitting the code into
parising and rendering.

The clean split should allow us to plug new renderers, say a Cairo
renderer to generate PDFs from the slides.