pinpoint: Generate PDFs
After the refactorig, we can acutally write a PDF renderer using cairo,
pango and gdk-pixbuf.

We really want the latest cairo to be able to use the neat features of
the PDF backend that allow you to generate small sized PDFs by always
using the same surface when painting the same image multiple times and
by embeding the jpeg directly when possible.