pinpoint: added custom transitions
When encountering a [transition=] command for a slide, pinpoint will
load the a file with the suffix .json in the current directory expecting
the following container skeleton structure:

"actor"
  "background"
  "midground"
  "foreground"

As well as a state machine with the states "pre", "show", "post" and
optionally "action". That animates this container skeleton (as well
as potential other actors).

See the transitions.txt presentation for more details.